Refactor: Package Structure Change
