+++ src/main/kotlin/com/example/ggyunispring/adapter/api/TestController.kt
@@ -1,4 +1,4 @@
-package com.example.ggyunispring.web
+package com.example.ggyunispring.adapter.api
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
