+++ src/main/kotlin/com/example/ggyunispring/adapter/dto/request/CreateDiaryRequestDTO.kt
@@ -1,4 +1,4 @@
-package com.example.ggyunispring.dto.request
+package com.example.ggyunispring.adapter.dto.request
 
 import com.example.ggyunispring.domain.enum.DiaryType
 import com.example.ggyunispring.domain.enum.Emotion
