+++ src/main/kotlin/com/example/ggyunispring/adapter/infrastructure/config/cors/WebMvcConfig.kt
@@ -1,11 +1,11 @@
-package com.example.ggyunispring.config.cors
+package com.example.ggyunispring.adapter.infrastructure.config.cors
 
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.servlet.config.annotation.CorsRegistry
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
 
 @Configuration
-class CorsConfig : WebMvcConfigurer {
+class WebMvcConfig : WebMvcConfigurer {
     override fun addCorsMappings(registry: CorsRegistry) {
         registry.addMapping("/api/v1/**")
             .allowedOriginPatterns("*")
