+++ src/main/kotlin/com/example/ggyunispring/adapter/infrastructure/config/mapper/ModelMapperConfig.kt
@@ -1,4 +1,4 @@
-package com.example.ggyunispring.config.mapper
+package com.example.ggyunispring.adapter.infrastructure.config.mapper
 
 import org.modelmapper.ModelMapper
 import org.springframework.context.annotation.Bean
