+++ src/main/kotlin/com/example/ggyunispring/application/DiaryService.kt
@@ -1,10 +1,10 @@
-package com.example.ggyunispring.service
+package com.example.ggyunispring.application
 
 import com.example.ggyunispring.domain.entity.Diary
 import com.example.ggyunispring.domain.repository.DiaryRepository
-import com.example.ggyunispring.dto.request.CreateDiaryRequestDTO
-import com.example.ggyunispring.dto.response.CreateDiaryResponseDTO
-import com.example.ggyunispring.dto.response.DiaryResponseDTO
+import com.example.ggyunispring.adapter.dto.request.CreateDiaryRequestDTO
+import com.example.ggyunispring.adapter.dto.response.CreateDiaryResponseDTO
+import com.example.ggyunispring.adapter.dto.response.DiaryResponseDTO
 import org.modelmapper.ModelMapper
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.security.core.userdetails.UserDetails
